Add ability to add source to charts
In order for a charts/viz to be taken serious, it must have the source
of the data attributed.

This PR add ability to define source either as a link via `data-source`
attribute on chart div (html) or as a `source` object in the `metadata`
section of chart data (python). The second method also offer ability to
separate the link and display title i.e.

```
html

<div id='' ...... data-source=''></div>

```

```
python

data['metadata']['source'] = {
    'title': 'Source Title',
    'link': 'https://link',
}

```